examples/tracing_middleware_compose/requirements.txt
git+https://github.com/i2gor87/panini.git@feature/tracing-middleware
opentelemetry-api==1.15.0
opentelemetry-sdk==1.15.0
opentelemetry-exporter-otlp-proto-grpc==1.15.0
opentelemetry-exporter-prometheus==1.12.0rc1
python-logging-loki==0.3.1